# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/article": {
      "get": {
        "operationId": "ArticleController_getArticles",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "페이지 번호",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지당 항목 수",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "정렬 기준: latest | likes",
            "schema": {
              "default": "latest",
              "example": "latest",
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "카테고리 이름 (선택)",
            "schema": {
              "example": "book",
              "type": "string",
              "enum": [
                "career",
                "job",
                "study",
                "tech",
                "lifestyle",
                "mentoring",
                "notice",
                "startup",
                "portfolio",
                "book",
                "etc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "아티클 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleListItemDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "서버 오류 또는 목록 조회 실패"
          }
        },
        "summary": "아티클 목록 조회",
        "tags": [
          "Article"
        ]
      },
      "post": {
        "operationId": "ArticleController_createArticle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateArticleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "아티클 작성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateArticleDto"
                }
              }
            }
          },
          "404": {
            "description": "멘토를 찾을 수 없습니다."
          },
          "500": {
            "description": "아티클 생성 중 오류가 발생했습니다."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "아티클 작성",
        "tags": [
          "Article"
        ]
      }
    },
    "/article/liked": {
      "get": {
        "operationId": "ArticleController_getLikedArticles",
        "parameters": [
          {
            "name": "ids",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "좋아요 여부확인",
        "tags": [
          "Article"
        ]
      }
    },
    "/article/{id}": {
      "get": {
        "operationId": "ArticleController_getArticleDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "아티클 상세 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleDetailDto"
                }
              }
            }
          }
        },
        "summary": "아티클 상세 조회 (조회수 증가)",
        "tags": [
          "Article"
        ]
      },
      "delete": {
        "operationId": "ArticleController_deleteArticle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "아티클 삭제 (작성자 전용)",
        "tags": [
          "Article"
        ]
      },
      "patch": {
        "operationId": "ArticleController_updateArticle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateArticleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "아티클 수정",
        "tags": [
          "Article"
        ]
      }
    },
    "/article/{id}/like": {
      "patch": {
        "operationId": "ArticleController_toggleArticleLike",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "좋아요 상태 토글 완료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToggleLikeResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "아티클이 존재하지 않음"
          },
          "500": {
            "description": "서버 오류"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "좋아요 토글",
        "tags": [
          "Article"
        ]
      }
    },
...
    "/article/{articleId}/comment": {
      "post": {
        "operationId": "ArticleController_createComment",
        "parameters": [
          {
            "name": "articleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "댓글/대댓글이 작성되었습니다."
          },
          "400": {
            "description": "잘못된 요청"
          },
          "404": {
            "description": "아티클을 찾을 수 없음"
          },
          "500": {
            "description": "서버 에러"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "댓글/대댓글 작성",
        "tags": [
          "Article"
        ]
      },
      "get": {
        "operationId": "ArticleController_getComment",
        "parameters": [
          {
            "name": "articleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "페이지 번호",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지당 항목 수",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "댓글 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentItemDto"
                }
              }
            }
          },
          "404": {
            "description": "아티클을 찾을 수 없음"
          },
          "500": {
            "description": "서버 에러"
          }
        },
        "summary": "댓글 목록 조회 (대댓글 포함)",
        "tags": [
          "Article"
        ]
      }
    },
    "/article/comment/{id}": {
      "patch": {
        "operationId": "ArticleController_updateComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "댓글이 수정되었습니다."
          },
          "400": {
            "description": "잘못된 요청"
          },
          "403": {
            "description": "수정 권한 없음"
          },
          "404": {
            "description": "댓글을 찾을 수 없음"
          },
          "500": {
            "description": "서버 에러"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "댓글 수정",
        "tags": [
          "Article"
        ]
      },
      "delete": {
        "operationId": "ArticleController_deleteComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "댓글이 삭제되었습니다."
          },
          "403": {
            "description": "삭제 권한 없음"
          },
          "404": {
            "description": "댓글을 찾을 수 없음"
          },
          "500": {
            "description": "서버 에러"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "댓글 삭제",
        "tags": [
          "Article"
        ]
      }
    },
...
  },
  "info": {
    "title": "커넥트 api문서",
    "description": "커넥트 개발을 위한 api문서 \n\n🔗 OpenAPI 명세 다운로드: /openapi-spec.json`,",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT Bearer Token"
      }
    },
    "schemas": {
...
      "ArticleListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-1234"
          },
          "title": {
            "type": "string",
            "example": "제목입니다.",
            "description": "아티클 제목"
          },
          "thumbnailUrl": {
            "type": "string",
            "example": "/uploads/article/thumbnail.jpg",
            "nullable": true
          },
          "views": {
            "type": "number",
            "example": 53
          },
          "likeCount": {
            "type": "number",
            "example": 12
          },
          "category": {
            "type": "string",
            "example": "book",
            "description": "아티클 카테고리",
            "enum": [
              "career",
              "job",
              "study",
              "tech",
              "lifestyle",
              "mentoring",
              "notice",
              "startup",
              "portfolio",
              "book",
              "etc"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "thumbnailUrl",
          "views",
          "likeCount",
          "category",
          "createdAt"
        ]
      },
      "CreateArticleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "제목입니다."
          },
          "content": {
            "type": "string",
            "example": "내용입니다."
          },
          "category": {
            "type": "string",
            "example": "book",
            "description": "아티클 카테고리",
            "enum": [
              "career",
              "job",
              "study",
              "tech",
              "lifestyle",
              "mentoring",
              "notice",
              "startup",
              "portfolio",
              "book",
              "etc"
            ]
          },
          "thumbnail": {
            "type": "string",
            "format": "binary",
            "description": "썸네일 이미지 파일"
          }
        },
        "required": [
          "title",
          "content",
          "category"
        ]
      },
      "ArticleDetailDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "아티클 제목입니다."
          },
          "content": {
            "type": "string",
            "example": "<p>아티클 내용입니다.</p><img src=\"...\">"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-07-08T15:30:00.000Z"
          },
          "views": {
            "type": "number",
            "example": 123,
            "description": "조회수"
          },
          "likeCount": {
            "type": "number",
            "example": 7,
            "description": "좋아요 수"
          },
          "authorNickname": {
            "type": "string",
            "example": "라이언",
            "description": "작성자 닉네임"
          },
          "authorImage": {
            "type": "string",
            "example": "https://cdn.konnect.com/profile.jpg",
            "description": "작성자 프로필 이미지 URL"
          },
          "authorId": {
            "type": "string",
            "example": "uuid-author-id",
            "description": "작성자 ID"
          }
        },
        "required": [
          "title",
          "content",
          "createdAt",
          "views",
          "likeCount",
          "authorNickname",
          "authorImage",
          "authorId"
        ]
      },
      "ToggleLikeResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "좋아요 추가됨"
          },
          "liked": {
            "type": "boolean",
            "example": true,
            "description": "true = 좋아요 됨, false = 좋아요 취소됨"
          }
        },
        "required": [
          "message",
          "liked"
        ]
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "좋은 글 감사합니다!"
          },
          "parentId": {
            "type": "string",
            "example": "uuid"
          }
        },
        "required": [
          "content"
        ]
      },
      "AuthorDto": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "example": "홍길동",
            "description": "닉네임"
          },
          "profile": {
            "type": "string",
            "example": "/uploads/profile.jpg",
            "description": "프로필 이미지"
          }
        },
        "required": [
          "nickname"
        ]
      },
      "ReplyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid",
            "description": "대댓글 ID"
          },
          "content": {
            "type": "string",
            "example": "저도 동의합니다!",
            "description": "대댓글 내용"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-09-10T12:30:00.000Z",
            "description": "작성일"
          },
          "author": {
            "$ref": "#/components/schemas/AuthorDto"
          }
        },
        "required": [
          "id",
          "content",
          "createdAt",
          "author"
        ]
      },
      "CommentItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid",
            "description": "댓글 ID"
          },
          "content": {
            "type": "string",
            "example": "좋은 글 감사합니다!",
            "description": "댓글 내용"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-09-10T12:00:00.000Z",
            "description": "작성일"
          },
          "author": {
            "$ref": "#/components/schemas/AuthorDto"
          },
          "children": {
            "description": "대댓글 목록",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReplyDto"
            }
          }
        },
        "required": [
          "id",
          "content",
          "createdAt",
          "author",
          "children"
        ]
      },
      "PatchCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "좋은 글 감사합니다!"
          }
        },
        "required": [
          "content"
        ]
      },
...
    }
  }
}